api/requirements.txt
pyrebase
flask
python-dotenv
flask-session
gunicorn